Add Python 3.9 to the build matrix
Apparently 3.10 and 3.11 are not (yet?) available on Travis?

$ curl -sSf --retry 5 -o python-3.10.tar.bz2 ${archive_url}
curl: (22) The requested URL returned error: 404 Not Found
Unable to download 3.10 archive. The archive may not exist. Please consider a different version.

$ curl -sSf --retry 5 -o python-3.11.tar.bz2 ${archive_url}
curl: (22) The requested URL returned error: 404 Not Found
Unable to download 3.11 archive. The archive may not exist. Please consider a different version.